requirements.txt
langchain
langchain-community
langchain-openai
openai
python-dotenv
slack-sdk 
slack-bolt 
Flask
gunicorn